Reworked swizzling to be way cleaner
