+++ MathSharp/Vector/Interfaces/IVec3.cs
@@ -6,28 +6,20 @@ namespace MathSharp
     /// <typeparam name="TBase">Base type of the vector. Must be of type <see cref="INumber{TSelf}"/>.</typeparam>
     /// <typeparam name="TFloat">Type used for forced float situations (such as <see cref="Mag"/>). Must be of type <see cref="IFloatingPoint{TSelf}"/>.</typeparam>
     /// <typeparam name="TVFloat">Type used for forced vector float situations (such as <see cref="Norm"/>). Must be of type <see cref="IVec3{TSelf, TBase, TFloat, TVFloat}"/>.</typeparam>
-    public interface IVec4<TSelf, TBase, TFloat, TVFloat>
+    public interface IVec3<TSelf, TBase, TFloat, TVFloat>
         where TSelf :
         struct,
-        IVec4<TSelf, TBase, TFloat, TVFloat>,
+        IVec3<TSelf, TBase, TFloat, TVFloat>,
         ISwizzlable<TSelf, TBase>,
         IEquatable<TSelf>
         where TBase : INumber<TBase>
         where TFloat : IFloatingPoint<TFloat>
         where TVFloat :
         struct,
-        IVec4<TVFloat, TFloat, TFloat, TVFloat>,
+        IVec3<TVFloat, TFloat, TFloat, TVFloat>,
         ISwizzlable<TVFloat, TFloat>,
         IEquatable<TVFloat>
     {
-        private static Dictionary<char, int> swizzleDict = new Dictionary<char, int>
-        {
-            { 'x', 0 },
-            { 'y', 1 },
-            { 'z', 2 },
-            { 'w', 3 }
-        };
-
         /// <summary>
         /// The x component of the vector.
         /// </summary>
@@ -44,11 +36,6 @@ namespace MathSharp
         TBase Z { get; set; }
 
         /// <summary>
-        /// The w component of the vector.
-        /// </summary>
-        TBase W { get; set; }
-
-        /// <summary>
         /// A new array representing the components of the vector.
         /// </summary>
         TBase[] Components { get; }
@@ -58,17 +45,14 @@ namespace MathSharp
         /// </summary>
         public TBase this[int i] { get; set; }
 
-        /// <summary>
-        /// For swizzling vectors. <see href="https://en.wikipedia.org/wiki/Swizzling_(computer_graphics)"/>.
-        /// I.e. Vec2 v = vec3["yz"];
-        /// </summary>
-        public Swizzle<TBase> this[string swizzle] { get; set; }
+        /// <inheritdoc cref="ISwizzlable{TSelf, TBase}.this[string]"/>
+        public TBase[] this[string swizzle] { get; set; }
 
-        //// <summary>
-        //// Computes the rotated vector by the given angle vector.
-        //// </summary>
-        //// <returns>A new floating point vector with the result of the rotation.</returns>
-        //public TVFloat Rotate(in AVec3 angle);
+        /// <summary>
+        /// Computes the rotated vector by the given angle vector.
+        /// </summary>
+        /// <returns>A new floating point vector with the result of the rotation.</returns>
+        public TVFloat Rotate(in AVec3 angle);
 
         /// <summary>
         /// Computes the squared magnitude of the vector.
@@ -86,6 +70,11 @@ namespace MathSharp
         public TBase Dot(in TSelf other);
 
         /// <summary>
+        /// Computes the cross product between two vectors. (<see href="https://en.wikipedia.org/wiki/Cross_product"/>).
+        /// </summary>
+        public TSelf Cross(in TSelf other);
+
+        /// <summary>
         /// Computes the normalized vector.
         /// </summary>
         public TVFloat Norm();
@@ -97,7 +86,7 @@ namespace MathSharp
 
         // Default implementations of required methods.
         /// <inheritdoc cref="Components"/>
-        public static TBase[] IComponents(in TSelf self) => new[] { self.X, self.Y, self.Z, self.W };
+        public static TBase[] IComponents(in TSelf self) => new[] { self.X, self.Y, self.Z };
 
         /// <inheritdoc cref="this[int]"/>
         public static TBase IIndexerGet(in TSelf self, int i)
@@ -110,8 +99,6 @@ namespace MathSharp
                     return self.Y;
                 case 2:
                     return self.Z;
-                case 3:
-                    return self.W;
                 default:
                     throw new IndexOutOfRangeException();
             }
@@ -131,43 +118,53 @@ namespace MathSharp
                 case 2:
                     self.Z = value;
                     break;
-                case 3:
-                    self.W = value;
-                    break;
                 default:
                     throw new IndexOutOfRangeException();
             }
         }
 
         /// <inheritdoc cref="this[string]"/>
-        public static Swizzle<TBase> ISwizzleGet(in TSelf self, string swizzle)
+        public static TBase[] ISwizzleGet(in TSelf self, string swizzle)
         {
-            return new Swizzle<TBase>(self.Components, swizzle);
-        }
-
-        /// <inheritdoc cref="this[string]"/>
-        public static void ISwizzleSet(
-            ref TSelf self,
-            string swizzleString,
-            Swizzle<TBase> swizzle,
-            Dictionary<char, int>? swizzleIndicesOverride = null)
-        {
-            Dictionary<char, int> swizzleIndices = swizzleIndicesOverride ?? swizzleDict;
-            if (swizzleString.Length > 4)
+            if (swizzle.Length > 3)
             {
-                throw new SwizzleException(4, swizzleString.Length);
+                throw new SwizzleException(3, swizzle.Length);
             }
 
-            if (swizzle.SwizzleString.Length != swizzleString.Length)
-            {
-                throw new SwizzleException(swizzleString.Length, swizzle.SwizzleString.Length);
-            }
-
-            for (int i = 0; i < swizzle.SwizzleString.Length; i++)
+            TBase[] result = new TBase[swizzle.Length];
+            for (int i = 0; i < swizzle.Length; i++)
             {
                 try
                 {
-                    self[swizzleIndices[swizzleString[i]]] = swizzle.Container[swizzleIndices[swizzle.SwizzleString[i]]];
+                    result[i] = self[TSelf.SwizzleMap[swizzle[i]]];
+                }
+                catch (KeyNotFoundException)
+                {
+                    throw new SwizzleException(swizzle[i]);
+                }
+            }
+
+            return result;
+        }
+
+        /// <inheritdoc cref="this[string]"/>
+        public static void ISwizzleSet(ref TSelf self, string swizzleString, TBase[] swizzle)
+        {
+            if (swizzleString.Length > 3)
+            {
+                throw new SwizzleException(3, swizzleString.Length);
+            }
+
+            if (swizzle.Length != swizzleString.Length)
+            {
+                throw new SwizzleMismatchException(swizzleString.Length, swizzle.Length);
+            }
+
+            for (int i = 0; i < swizzle.Length; i++)
+            {
+                try
+                {
+                    self[TSelf.SwizzleMap[swizzleString[i]]] = swizzle[i];
                 }
                 catch (KeyNotFoundException)
                 {
@@ -176,58 +173,42 @@ namespace MathSharp
             }
         }
 
-        /// <inheritdoc cref="ISwizzlable{TSelf, TBase}.implicit operator TSelf(Swizzle{TBase})"/>
-        public static TSelf ISwizzleToSelf(
-            Swizzle<TBase> swizzle,
-            Dictionary<char, int>? swizzleIndicesOverride = null)
+        /// <inheritdoc cref="ISwizzlable{TSelf, TBase}.implicit operator TSelf"/>
+        public static TSelf ISwizzleToSelf(TBase[] swizzle)
         {
-            Dictionary<char, int> swizzleIndices = swizzleIndicesOverride ?? swizzleDict;
-            if (swizzle.SwizzleString.Length != 4)
+            if (swizzle.Length != 3)
             {
-                throw new SwizzleException(4, swizzle.SwizzleString.Length);
+                throw new SwizzleMismatchException(3, swizzle.Length);
             }
 
-            TSelf result = new TSelf { };
-            for (int i = 0; i < swizzle.SwizzleString.Length; i++)
-            {
-                try
-                {
-                    result[i] = swizzle.Container[swizzleIndices[swizzle.SwizzleString[i]]];
-                }
-                catch (KeyNotFoundException)
-                {
-                    throw new SwizzleException(swizzle.SwizzleString[i]);
-                }
-            }
-
-            return result;
+            return new TSelf { X = swizzle[0], Y = swizzle[1], Z = swizzle[2] };
         }
 
-        /// <inheritdoc cref="ISwizzlable{TSelf, TBase}.implicit operator Swizzle{TBase}"/>
-        public static Swizzle<TBase> ISelfToSwizzle(in TSelf self, string fullSwizzle)
+        /// <inheritdoc cref="ISwizzlable{TSelf, TBase}.implicit operator TBase[](in TSelf)"/>
+        public static TBase[] ISelfToSwizzle(in TSelf self)
         {
-            return new Swizzle<TBase>(self.Components, fullSwizzle);
+            return self.Components;
         }
 
-        /// <summary>
-        /// Creates a swizzle object that can be used to swizzle vectors. <see href="https://en.wikipedia.org/wiki/Swizzling_(computer_graphics)"/>.
-        /// </summary>
-        public static Swizzle<TBase> ISwizzle(in TSelf self, string swizzle)
-        {
-            return new Swizzle<TBase>(self.Components, swizzle);
-        }
-
-        //// <inheritdoc cref="Rotate(in AVec3)"/>
-        //public TVFloat IRotate(in AVec3 angle) => throw new NotImplementedException();
+        /// <inheritdoc cref="Rotate(in AVec3)"/>
+        public static TVFloat IRotate(in TSelf self, in AVec3 angle) => throw new NotImplementedException();
 
         /// <inheritdoc cref="Mag2"/>
-        public static TBase IMag2(in TSelf self) => self.X * self.X + self.Y * self.Y + self.Z * self.Z + self.W * self.W;
+        public static TBase IMag2(in TSelf self) => self.X * self.X + self.Y * self.Y + self.Z * self.Z;
 
         /// <inheritdoc cref="Mag"/>
         public static TFloat IMag(in TSelf self) => ToTFloat(Math.Sqrt(Convert.ToDouble(self.Mag2())));
 
         /// <inheritdoc cref="Dot"/>
-        public static TBase IDot(in TSelf self, in TSelf other) => self.X * other.X + self.Y * other.Y + self.Z * other.Z + self.W * other.W;
+        public static TBase IDot(in TSelf self, in TSelf other) => self.X * other.X + self.Y * other.Y + self.Z * other.Z;
+
+        /// <inheritdoc cref="Cross(in TSelf)"/>
+        public static TSelf ICross(in TSelf self, in TSelf other) => new TSelf
+        {
+            X = self.Y * other.Z - self.Z * other.Y,
+            Y = self.Z * other.X - self.X * other.Z,
+            Z = self.X * other.Y - self.Y * other.X
+        };
 
         /// <inheritdoc cref="Norm"/>
         public static TVFloat INorm(in TSelf self) => IFDiv(self, self.Mag());
@@ -236,50 +217,50 @@ namespace MathSharp
         /// Computes the sum of two vectors.
         /// </summary>
         public static TSelf IAdd(in TSelf self, in TSelf other)
-            => new TSelf { X = self.X + other.X, Y = self.Y + other.Y, Z = self.Z + other.Z, W = self.W + other.W };
+            => new TSelf { X = self.X + other.X, Y = self.Y + other.Y, Z = self.Z + other.Z };
 
         /// <summary>
         /// Computes the difference of two vectors.
         /// </summary>
         public static TSelf ISub(in TSelf self, in TSelf other)
-            => new TSelf { X = self.X - other.X, Y = self.Y - other.Y, Z = self.Z - other.Y, W = self.W - other.W };
+            => new TSelf { X = self.X - other.X, Y = self.Y - other.Y, Z = self.Z - other.Y };
 
         /// <summary>
         /// Computes the Hadamard product of two vectors, also known as the component-wise product (<see href="https://en.wikipedia.org/wiki/Hadamard_product_(matrices)"/>).
         /// </summary>
         public static TSelf IMul(in TSelf self, in TSelf other)
-            => new TSelf { X = self.X * other.X, Y = self.Y * other.Y, Z = self.Z * other.Z, W = self.W * other.W };
+            => new TSelf { X = self.X * other.X, Y = self.Y * other.Y, Z = self.Z * other.Z };
 
         /// <summary>
         /// Computes the product of a vector and a scalar.
         /// </summary>
         public static TSelf IMul(in TSelf self, TBase scalar)
-            => new TSelf { X = self.X * scalar, Y = self.Y * scalar, Z = self.Z * scalar, W = self.W * scalar };
+            => new TSelf { X = self.X * scalar, Y = self.Y * scalar, Z = self.Z * scalar };
 
         /// <inheritdoc cref="IMul(in TSelf, TBase)"/>
         public static TVFloat IFMul(in TSelf self, TFloat scalar)
-            => new TVFloat { X = ToTFloat(self.X) * scalar, Y = ToTFloat(self.Y) * scalar, Z = ToTFloat(self.Z) * scalar, W = ToTFloat(self.W) * scalar };
+            => new TVFloat { X = ToTFloat(self.X) * scalar, Y = ToTFloat(self.Y) * scalar, Z = ToTFloat(self.Z) * scalar };
 
         /// <summary>
         /// Computes the Hadamar inverse product (division) of two vectors, also known as the component-wise inverse product (<see href="https://en.wikipedia.org/wiki/Hadamard_product_(matrices)"/>).
         /// </summary>
         public static TSelf IDiv(in TSelf self, in TSelf other)
-            => new TSelf { X = self.X / other.X, Y = self.Y / other.Y, Z = self.Z / other.Z, W = self.W / other.W };
+            => new TSelf { X = self.X / other.X, Y = self.Y / other.Y, Z = self.Z / other.Z };
 
         /// <summary>
         /// Computes the division of a vector by a scalar.
         /// </summary>
         public static TSelf IDiv(in TSelf self, TBase scalar)
-            => new TSelf { X = self.X / scalar, Y = self.Y / scalar, Z = self.Z / scalar, W = self.W / scalar };
+            => new TSelf { X = self.X / scalar, Y = self.Y / scalar, Z = self.Z / scalar };
 
         /// <inheritdoc cref="IDiv(in TSelf, TBase)"/>
         public static TVFloat IFDiv(in TSelf self, TFloat scalar)
-            => new TVFloat { X = ToTFloat(self.X) / scalar, Y = ToTFloat(self.Y) / scalar, Z = ToTFloat(self.Z) / scalar, W = ToTFloat(self.W) / scalar };
+            => new TVFloat { X = ToTFloat(self.X) / scalar, Y = ToTFloat(self.Y) / scalar, Z = ToTFloat(self.Z) / scalar };
 
         /// <summary>
         /// Computes whether two vectors are equal.
         /// </summary>
-        public static bool IEquals(in TSelf self, in TSelf other) => self.X == other.X && self.Y == other.Y && self.Z == other.Z && self.W == other.W;
+        public static bool IEquals(in TSelf self, in TSelf other) => self.X == other.X && self.Y == other.Y && self.Z == other.Z;
 
         /// <inheritdoc cref="object.Equals(object?)"/>
         public static bool IEquals(in TSelf self, in object? obj) => obj is TSelf other && IEquals(self, other);
@@ -288,7 +269,7 @@ namespace MathSharp
         public static int IGetHashCode(in TSelf self) => ((object)self).GetHashCode();
 
         /// <inheritdoc cref="ToString"/>
-        public static string IToString(in TSelf self) => $"<{self.X}, {self.Y}, {self.Z}, {self.W}>";
+        public static string IToString(in TSelf self) => $"<{self.X}, {self.Y}, {self.Z}>";
 
         private static TFloat ToTFloat(double d) => (TFloat)Convert.ChangeType(d, typeof(TFloat));
