+++ MathSharp/Vector/Swizzling/ISwizzlable.cs
@@ -1,5 +1,8 @@
 ﻿namespace MathSharp
 {
+    // Ideas:
+    // Swizzling creates an array instead of a new type. Remove the Swizzle class entirely and use an array instead.
+    // Swizzle interface
     /// <summary>
     /// Interface for swizzlable types.
     /// </summary>
@@ -7,18 +10,24 @@
         where TSelf : struct, ISwizzlable<TSelf, TBase>
     {
         /// <summary>
+        /// Mapping of swizzle characters to their index in the implementing type.
+        /// i.e. SwizzleMap['x'] = 0, SwizzleMap['y'] = 1, etc.
+        /// </summary>
+        public static abstract Dictionary<char, int> SwizzleMap { get; }
+
+        /// <summary>
         /// For implicitly converting a swizzle to the implementing type. i.e. <c>Vec2 v = vec3["yz"];</c>
         /// </summary>
-        public static abstract implicit operator TSelf(Swizzle<TBase> swizzler);
+        public static abstract implicit operator TSelf(TBase[] swizzle);
 
         /// <summary>
         /// For implicitly converting the implementing type to a swizzle. i.e. <c>Vec2 v["yx"] = other2;</c>
         /// </summary>
-        public static abstract implicit operator Swizzle<TBase>(in TSelf swizzler);
+        public static abstract implicit operator TBase[](in TSelf swizzler);
 
         /// <summary>
         /// Used to perform a swizzle. i.e. <c>vec3["yz"] = vec3["xz"];</c>
         /// </summary>
-        public Swizzle<TBase> this[string swizzle] { get; set; }
+        public TBase[] this[string swizzle] { get; set; }
     }
 }
